rag_from_scratch/requirements.txt
# Core Dependencies
faiss-cpu>=1.7.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.0.0
numpy>=1.20.0

# Optional: Local LLM support
ollama>=0.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
